requirements.txt
psycopg2-binary==2.9.7
schedule==1.2.0
pysnmp==4.4.12
pyasn1==0.4.8
python-dotenv==1.0.0
pytest==7.4.0
pyflakes==2.6.0